requirements.txt
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
html5lib==1.1
pandas==2.2.3
httpx==0.24.1
aiohttp==3.8.5
urlextract==1.9.0
python-dateutil==2.8.2
tqdm==4.66.1
playwright==1.41.1  # optional - install and run `playwright install` if needed
